아이돌/소속사 등록

+++ user-service/src/main/java/com/bit/idol/userservice/controller/AgencyController.java
@@ -36,7 +36,7 @@ public class AgencyController {
     }
 
     @GetMapping("/{agencyId}")
-    public ResponseEntity<AgencyDto> getAgency(@PathVariable Long agencyId) {
+    public ResponseEntity<AgencyDto> getAgency(@PathVariable int agencyId) {
         return ResponseEntity.ok(agencyService.getAgency(agencyId));
     }
