+++ user-service/src/main/java/com/bit/idol/userservice/service/AgencyService.java
@@ -31,7 +31,7 @@ public class AgencyService {
         return AgencyDto.fromEntity(agency);
     }
 
-    public AgencyDto getAgency(Long agencyId) {
+    public AgencyDto getAgency(int agencyId) {
         Agency agency = agencyRepository.findById(agencyId)
                 .orElseThrow(() -> new RuntimeException("Agency not found"));
         return AgencyDto.fromEntity(agency);
